+++ Online_Shopping/Controllers/BranchController.cs
@@ -9,6 +9,7 @@ using Repositories.Repositories;
 using Repository.Contracts.Interfaces;
 using Service.Contracts.Interfaces;
 using Services.Services;
+using System.Net.Http;
 
 namespace Online_Shopping.Controllers
 {
@@ -18,13 +19,18 @@ namespace Online_Shopping.Controllers
     {
         private readonly IBranchService _branchService;
         private readonly IMapper _mapper;
+        private readonly HttpClient _httpClient;
         private readonly IAddressRepo _addressRepo;
 
+        private readonly string api = "http://localhost:5285/api/branches/north";
+
         public BranchController(IBranchService branchService, IMapper mapper,
+            HttpClient httpClient,
             IAddressRepo addressRepo)
         {
             _branchService = branchService;
             _mapper = mapper;
+            _httpClient = httpClient;
             _addressRepo = addressRepo;
         }
 
@@ -32,6 +38,7 @@ namespace Online_Shopping.Controllers
         public async Task<IActionResult> AddNewBranch([FromBody] RequestBranch requestBranch)
         {
             BranchDTO branchDTO = await _branchService.AddNewBranch(requestBranch);
+            await _httpClient.PostAsJsonAsync($"{api}/new-branch/{branchDTO.Id}", requestBranch);
 
             return CreatedAtAction(nameof(GetAllBranches), new { id = branchDTO.Id }, requestBranch);
         }
@@ -40,8 +47,14 @@ namespace Online_Shopping.Controllers
         public async Task<IActionResult> GetAllBranches()
         {
             var list = await _branchService.GetBranchList();
-            if (list == null)
-                return NotFound();
+            if (list.Count == 0)
+            {
+                return NotFound(new Response<List<BranchDTO>>
+                {
+                    Message = "Don't have any branches!",
+                    Data = list
+                });
+            }
             return Ok(list);
         }
 
@@ -49,22 +62,68 @@ namespace Online_Shopping.Controllers
         public async Task<IActionResult> GetBranch(string id)
         {
             var branch = await _branchService.GetBranch(id);
+            
             if (branch == null)
-                return NotFound();
+            {
+                return NotFound(new Response<string>
+                {
+                    Message = "This branch does not exist!"
+                });
+            }
+
+            if (branch.RegionId == "Bac")
+            {
+                var response = await _httpClient.GetAsync($"{api}/{id}");
+                if (response.IsSuccessStatusCode)
+                {
+                    return Ok(await response.Content.ReadFromJsonAsync<BranchDTO>());
+                }
+            }
+
             return Ok(branch);
         }
 
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteBranch(string id)
         {
-            await _branchService.DeleteBranch(id);
+            var branch = await _branchService.GetBranch(id);
+            if (branch == null)
+            {
+                return NotFound(new Response<string>
+                {
+                    Message = "This branch does not exist!"
+                });
+            }
+            else
+            {
+                await _branchService.DeleteBranch(id);
+                if (branch.RegionId == "Bac")
+                {
+                    await _httpClient.DeleteAsync($"{api}/{id}");
+                }
+            }
+            
             return NoContent();
         }
 
         [HttpPut("{id}")]
         public async Task<IActionResult> UpdateBranch(string id, [FromBody] RequestBranch requestBranch)
         {
+            if (await _branchService.GetBranch(id) == null)
+            {
+                return NotFound(new Response<string>
+                {
+                    Message = "This branch does not exist!"
+                });
+            }
+
             await _branchService.UpdateBranch(id, requestBranch);
+
+            if (requestBranch.RegionId == "Bac")
+            {
+                await _httpClient.PutAsJsonAsync($"{api}/{id}", requestBranch);
+            }
+
             return NoContent();
         }
     }
